refactor: move the mutation observer to the provider
